feat: defined radii
