+++ components/00-tokens/effects/radii.twig
@@ -1,0 +1,19 @@
+{% set radii__base_class = "cl-radii" %}
+
+<div {{ bem(radii__base_class) }}>
+  <ul {{ bem('list', [], radii__base_class) }} data-theme="">
+  {% for radius, value in _context.radii %}
+    {% set variable = prefix ~ radius|lower|replace({' ': '-'}) %}
+      <li {{ bem("item", [], radii__base_class) }} style="border-radius: var({{ variable }})">
+        <span {{ bem("title", [], radii__base_class) }}>Radius {{ radius }}</span><br />
+      <details class="cl-info">
+        <summary>Radius Details</summary>
+        <code>
+          <strong>variable:</strong> {{ variable }} <br />
+          <strong>value:</strong> {{ value }}
+        </code>
+      </details>
+      </li>
+    {% endfor %}
+  </ul>
+</div>
